+++ GetThePicture.Tests/Picture/Clause/PicClauseCodec-Comp5Test.cs
@@ -5,7 +5,7 @@ using GetThePicture.Picture.Clause.Base.ClauseItems;
 namespace GetThePicture.Tests.Picture.Clause;
 
 [TestClass]
-public class CompTest
+public class CodecComp5Test
 {
     // -------------------------
     // Encode
@@ -24,9 +24,9 @@ public class CompTest
     public void Encode_Combination_Test(string picString, object value, byte[] expected)
     {
         var pic = PicMeta.Parse(picString);
-        pic.Usage = PicUsage.Binary;
+        pic.Usage = PicUsage.NativeBinary;
 
-        byte[] buffer = PicClauseCodec.ForMeta(pic).WithStrict().Encode(value);
+        byte[] buffer = PicClauseCodec.ForMeta(pic).WithStrict().IsLittleEndian().Encode(value);
 
         CollectionAssert.AreEqual(expected, buffer);
     }
@@ -37,9 +37,9 @@ public class CompTest
     public void Encode_With_Change_Endian(string picString, object value, byte[] expected)
     {
         var pic = PicMeta.Parse(picString);
-        pic.Usage = PicUsage.Binary;
+        pic.Usage = PicUsage.NativeBinary;
 
-        byte[] buffer = PicClauseCodec.ForMeta(pic).WithStrict().WithReversedBinary().Encode(value);
+        byte[] buffer = PicClauseCodec.ForMeta(pic).WithStrict().Encode(value);
 
         CollectionAssert.AreEqual(expected, buffer);
     }
@@ -61,9 +61,9 @@ public class CompTest
     public void Decode_Combination_Test(string picString, object expected, byte[] buffer)
     {
         var pic = PicMeta.Parse(picString);
-        pic.Usage = PicUsage.Binary;
+        pic.Usage = PicUsage.NativeBinary;
 
-        var value = PicClauseCodec.ForMeta(pic).WithStrict().Decode(buffer);
+        var value = PicClauseCodec.ForMeta(pic).WithStrict().IsLittleEndian().Decode(buffer);
 
         Assert.AreEqual(expected, value);
     }
@@ -74,9 +74,9 @@ public class CompTest
     public void Decode_With_Change_Endian(string picString, object expected, byte[] buffer)
     {
         var pic = PicMeta.Parse(picString);
-        pic.Usage = PicUsage.Binary;
+        pic.Usage = PicUsage.NativeBinary;
 
-        var value = PicClauseCodec.ForMeta(pic).WithStrict().WithReversedBinary().Decode(buffer);
+        var value = PicClauseCodec.ForMeta(pic).WithStrict().Decode(buffer);
 
         Assert.AreEqual(expected, value);
     }
@@ -89,10 +89,23 @@ public class CompTest
     [ExpectedException(typeof(OverflowException))]
     public void Encode_Without_Sign_Negative_ThrowsOverflowException()
     {
-        // PIC 9(4) COMP.
+        // PIC 9(4) COMP-4.
         var pic = PicMeta.Parse("9(4)");
         pic.Usage = PicUsage.Binary;
 
         PicClauseCodec.ForMeta(pic).WithStrict().Encode(-9999);
+    }
+
+    [DataTestMethod]
+    [DataRow( "9(28)",  9999)]
+    [DataRow("S9(28)",  9999)]
+    [DataRow("S9(28)", -9999)]
+    [ExpectedException(typeof(NotSupportedException))]
+    public void Encode_For_BigInt_NotSupported(string picString, object value)
+    {
+        var pic = PicMeta.Parse(picString);
+        pic.Usage = PicUsage.Binary;
+
+        PicClauseCodec.ForMeta(pic).WithStrict().Encode(value);
     }
 }