+++ GetThePicture/Picture/Clause/Codec/Category/Numeric/Decoder.cs
@@ -1,6 +1,7 @@
 using GetThePicture.Picture.Clause.Base;
 using GetThePicture.Picture.Clause.Base.ClauseItems;
 using GetThePicture.Picture.Clause.Base.Computational;
+using GetThePicture.Picture.Clause.Base.Computational.Base;
 using GetThePicture.Picture.Clause.Base.Options;
 
 using GetThePicture.Picture.Clause.Codec.Category.Numeric.Mapper;
@@ -32,11 +33,11 @@ public static class Decoder
 
         return pic.Usage switch
         {
-            PicUsage.Display        => Display_Decode(bytes, pic, options),
-            PicUsage.PackedDecimal  =>   COMP3.Decode(bytes, pic),
-            PicUsage.Binary         =>   COMP4.Decode(bytes, pic, options.Binary),
-            PicUsage.NativeBinary   =>   COMP5.Decode(bytes, pic, options.Binary),
-            PicUsage.UPackedDecimal =>   COMP6.Decode(bytes, pic),
+            PicUsage.Display => Display_Decode(bytes, pic, options),
+            PicUsage.COMP3   =>   COMP3.Decode(bytes, pic),
+            PicUsage.COMP4   =>   COMP4.Decode(bytes, pic),
+            PicUsage.COMP5   =>   COMP5.Decode(bytes, pic, options.IsBigEndian),
+            PicUsage.COMP6   =>   COMP6.Decode(bytes, pic),
             _ => throw new NotSupportedException($"Unsupported numeric storage: {pic.Usage}")
         };
     }