+++ GetThePicture/Picture/Clause/Codec/Category/Numeric/Encoder.cs
@@ -27,11 +27,11 @@ public static class Encoder
 
         byte[] buffer = pic.Usage switch
         {
-            PicUsage.Display        => Display_Encode(nMeta, pic, options),
-            PicUsage.PackedDecimal  =>   COMP3.Encode(nMeta, pic),
-            PicUsage.Binary         =>   COMP4.Encode(nMeta, pic, options.Binary),
-            PicUsage.NativeBinary   =>   COMP5.Encode(nMeta, pic, options.Binary),
-            PicUsage.UPackedDecimal =>   COMP6.Encode(nMeta, pic),
+            PicUsage.Display => Display_Encode(nMeta, pic, options),
+            PicUsage.COMP3   =>   COMP3.Encode(nMeta, pic),
+            PicUsage.COMP4   =>   COMP4.Encode(nMeta, pic),
+            PicUsage.COMP5   =>   COMP5.Encode(nMeta, pic, options.IsBigEndian),
+            PicUsage.COMP6   =>   COMP6.Encode(nMeta, pic),
             _ => throw new NotSupportedException($"Unsupported numeric storage: {pic.Usage}")
         };